Move mongo client tracing to opentelemetry
As suggested by @brunobat, moving the implementation
from mongo-client to opentelemetry extension
