+++ extensions/opentelemetry/runtime/src/test/java/io/quarkus/opentelemetry/runtime/tracing/intrumentation/mongodb/MongoTracingCommandListenerTest.java
@@ -1,4 +1,4 @@
-package io.quarkus.mongodb.tracing;
+package io.quarkus.opentelemetry.runtime.tracing.intrumentation.mongodb;
 
 import static org.assertj.core.api.Assertions.assertThatNoException;
